% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_tibble}
\alias{empty_tibble}
\title{Create empty tibble}
\usage{
empty_tibble(column_names, column_types)
}
\arguments{
\item{column_names}{character vector with column names}

\item{column_types}{character vector with column types}
}
\value{
empty tibble object
}
\description{
This functions creates empty tibble with columns names specified
as parameter.
}
\examples{
empty_tibble(c("a", "b"), c("character", "integer"))
}
