% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{get_json}
\alias{get_json}
\title{Tipper Raw JSON}
\usage{
get_json(tipper_id)
}
\arguments{
\item{tipper_id}{The id of the tipper you want.}
}
\value{
a \code{list}
}
\description{
Returns raw json from specific action network tipper
}
\examples{
get_json(tipper_id = "184328")
}
