% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{get_stats_table}
\alias{get_stats_table}
\title{Tipper Stats Table}
\usage{
get_stats_table(json)
}
\arguments{
\item{json}{The json data of the tipper you want stats, obtained
by \code{get_json} function call.}
}
\value{
a \code{tibble} with 6 columns
\describe{
\item{tipper_id}{unique id of the tipper}
\item{league}{league name of the tip stats}
\item{period}{period of sampling for calculate stats}
\item{win}{count of wins in the specif league and time period}
\item{loss}{count of losses in the specif league and time period}
\item{count}{count of tips in the specif league and time period}
}
}
\description{
Returns a tibble with id, league, periods and W/L of tips
}
\examples{
json <- get_json(tipper_id = "184328")
get_stats_table(json = json)
}
