% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{get_tipper_table}
\alias{get_tipper_table}
\title{Tipper Table Summary}
\usage{
get_tipper_table(json)
}
\arguments{
\item{json}{The json data of the tipper you want information, obtained
by \code{get_json} function call.}
}
\value{
a \code{tibble} with 6 columns
\describe{
\item{tipper_id}{unique id of the tipper}
\item{name}{name of the tipper}
\item{is_expert}{the tipper is an expert?}
\item{is_author}{the tipper is an author of sport articles?}
\item{is_verified}{the tipper is verified/famous in action network?}
\item{num_followers}{total number of followers in action network}
}
}
\description{
Returns a tibble with id, name and others informations
}
\examples{
json <- get_json(tipper_id = "184328")
get_tipper_table(json = json)
}
