% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{get_tippers_table}
\alias{get_tippers_table}
\title{Tippers Table Summary}
\usage{
get_tippers_table(tippers_id)
}
\arguments{
\item{tippers_id}{A vector with all tippers id's you want information.}
}
\value{
a \code{tibble} with 6 columns
\describe{
\item{tipper_id}{unique id of the tipper}
\item{name}{name of the tipper}
\item{is_expert}{the tipper is an expert?}
\item{is_author}{the tipper is an author of sport articles?}
\item{is_verified}{the tipper is verified/famous in action network?}
\item{num_followers}{total number of followers in action network}
}
}
\description{
Returns a tibble with id, name and others informations of each
tipper requested
}
\examples{
get_tippers_table(tippers_id = c(184328, 184329))
}
