% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{get_tips_table}
\alias{get_tips_table}
\title{Tippers Tips Table}
\usage{
get_tips_table(tippers_id)
}
\arguments{
\item{tippers_id}{A vector with all tippers id's you want tips.}
}
\value{
a \code{tibble} with 6 columns
\describe{
\item{tip_id}{unique id of the tip}
\item{tipper_id}{unique id of the tipper}
\item{game_id}{unique id of the game where the tip is on}
\item{created_at}{date where the tip has created}
\item{updated_at}{date where the tip has updated}
\item{league}{league name of the game}
\item{tip_play}{the tip play}
\item{tip_type}{the tip type}
}
}
\description{
Returns a tibble with id, league, periods and W/L of tips
}
\examples{
get_tips_table(tippers_id = c(184328, 184329))
}
