% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{get_stats_table}
\alias{get_stats_table}
\title{Tippers Stats Table}
\usage{
get_stats_table(tippers_id = NA, verbose = FALSE)
}
\arguments{
\item{tippers_id}{A vector with all tippers id's you want stats.}

\item{verbose}{Logical indication showing scrapping progress bar}
}
\value{
a \code{tibble} with 6 columns
\describe{
\item{tipper_id}{unique id of the tipper}
\item{league}{league name of the tip stats}
\item{period}{period of sampling for calculate stats}
\item{win}{count of wins in the specif league and time period}
\item{loss}{count of losses in the specif league and time period}
\item{count}{count of tips in the specif league and time period}
}
}
\description{
Returns a tibble with id, league, periods and W/L of tips
}
\examples{
get_stats_table(tippers_id = c(184328, 184329))
}
