% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{get_tips_table}
\alias{get_tips_table}
\title{Tipper Tips Table}
\usage{
get_tips_table(json)
}
\arguments{
\item{json}{The json data of the tipper you want tips, obtained
by \code{get_json} function call.}
}
\value{
a \code{tibble} with 6 columns
\describe{
\item{pick_id}{unique id of the tip}
\item{tipper_id}{unique id of the tipper}
\item{game_id}{unique id of the game where the tip is on}
\item{created_at}{date where the tip has created}
\item{updated_at}{date where the tip has updated}
\item{league}{league name of the game}
\item{tip_play}{the tip play}
\item{tip_type}{the tip type}
}
}
\description{
Returns a tibble with id, league, periods and W/L of tips
}
\examples{
json <- get_json(tipper_id = "184328")
get_stats_table(json = json)
}
